Add standard dynamic behavior to _sikuli_runall
Sikuli_runall needed to be edited in order to get add tests.
This new way of using sikuli_runall creates setup and teardown methods that
you can put the _sikuli_runall folder and the __init__.py and optionally
setUp and tearDown files in a folder of tests and the tests will be added
and run with "runsikulix(.cmd) -r _sikuli_runall".
It will be easier to maintain this than having to set up a test class for
each folder.