Add one-liner ability to open fwbackup or fwdata
To open a flex project or backup one would use the file menu of a preopened
flex project.
Fwbackup files are now unzipped and the fwdata file is opened. For an
fwdata file a subprocess is started to open the file.
This will make it more direct to open a flex project.